+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.12.0]
+### Added
+- Add tooltip if option label in `Select` overflows
+
+### Refactored
+- Set text color in input components default to `$base-120` and `$base-60` when they are disabled
+
 ## [0.11.10]
 ### Refactored
 - Change type `tel` to `float` in `type` prop of `ValidFieldText`